+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/includes/autowiring_classes_80.php
@@ -2,6 +2,7 @@
 
 namespace Symfony\Component\DependencyInjection\Tests\Compiler;
 
+use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Contracts\Service\Attribute\Required;
 
 class AutowireSetter
@@ -25,4 +26,19 @@ class AutowireProperty
 {
     #[Required]
     public Foo $foo;
+}
+
+class AutowireAttribute
+{
+    public function __construct(
+        #[Autowire(service: 'some.id')]
+        public \stdClass $service,
+        #[Autowire(expression: "parameter('some.parameter')")]
+        public string $expression,
+        #[Autowire(value: '%some.parameter%/bar')]
+        public string $value,
+        #[Autowire(service: 'invalid.id')]
+        public ?\stdClass $invalid = null,
+    ) {
+    }
 }