+++ src/Symfony/Component/HttpKernel/Tests/DependencyInjection/RegisterControllerArgumentLocatorsPassTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\HttpKernel\Tests\DependencyInjection;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\Argument\ServiceClosureArgument;
+use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\DependencyInjection\Attribute\Target;
 use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
@@ -434,6 +435,36 @@ class RegisterControllerArgumentLocatorsPassTest extends TestCase
         $expected = ['apiKey' => new ServiceClosureArgument(new Reference('the_api_key'))];
         $this->assertEquals($expected, $locator->getArgument(0));
     }
+
+    public function testAutowireAttribute()
+    {
+        if (!class_exists(Autowire::class)) {
+            $this->markTestSkipped('#[Autowire] attribute not available.');
+        }
+
+        $container = new ContainerBuilder();
+        $resolver = $container->register('argument_resolver.service', 'stdClass')->addArgument([]);
+
+        $container->register('some.id', \stdClass::class);
+        $container->setParameter('some.parameter', 'foo');
+
+        $container->register('foo', WithAutowireAttribute::class)
+            ->addTag('controller.service_arguments');
+
+        (new RegisterControllerArgumentLocatorsPass())->process($container);
+
+        $locatorId = (string) $resolver->getArgument(0);
+        $container->getDefinition($locatorId)->setPublic(true);
+
+        $container->compile();
+
+        $locator = $container->get($locatorId)->get('foo::fooAction');
+
+        $this->assertInstanceOf(\stdClass::class, $locator->get('service1'));
+        $this->assertSame('foo/bar', $locator->get('value'));
+        $this->assertSame('foo', $locator->get('expression'));
+        $this->assertFalse($locator->has('service2'));
+    }
 }
 
 class RegisterTestController
@@ -508,6 +539,21 @@ class WithTarget
     public function fooAction(
         #[Target('some.api.key')]
         string $apiKey
+    ) {
+    }
+}
+
+class WithAutowireAttribute
+{
+    public function fooAction(
+        #[Autowire(service: 'some.id')]
+        \stdClass $service1,
+        #[Autowire(value: '%some.parameter%/bar')]
+        string $value,
+        #[Autowire(expression: "parameter('some.parameter')")]
+        string $expression,
+        #[Autowire(service: 'invalid.id')]
+        \stdClass $service2 = null,
     ) {
     }
 }